refactor(data): split webdataset storage modules
Changes:
- Move WebDataset index, packing, streaming-load, and shared filesystem/tar concerns into `datasets.webdataset`.
- Remove the unmerged `webdataset_io` module and update CLI, training, dataset, documentation, and tests to use direct owners.

Impact:
- Maintainers can navigate four focused modules without retaining a legacy compatibility surface.
- Packing and streaming behavior remain available through the dedicated CLI and dataset construction paths.

Verification:
- `pre-commit run --all-files` passed.
- 135 CLI, COCO integration, and moved-module doctests passed; 33 focused loader/build tests and the package-layout test passed.

Residual limits:
- The complete 142-test streaming module did not produce a terminal capture in this environment; rerun it in CI or an interactive terminal.

---
