+++ src/rfdetr/config.py
@@ -1058,7 +1058,7 @@ class TrainConfig(BaseConfig):
     keypoint_nll_loss_coef: float = 0
     keypoint_oks_sigmas: list[float] | None = None
     # "webdataset" streams pre-packed tar shards instead of loose image files; see
-    # rfdetr.datasets.webdataset_io for the packer and the sizing contract it imposes on the loaders.
+    # rfdetr.datasets.webdataset for the packer and the sizing contract it imposes on the loaders.
     dataset_file: Literal["coco", "o365", "roboflow", "yolo", "webdataset"] = "roboflow"
     square_resize_div_64: bool = True
     dataset_dir: PathLikeStr | None
