+++ src/rfdetr/datasets/webdataset/load.py
@@ -3,55 +3,27 @@
 # Copyright (c) 2025 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-"""Sequential-I/O dataset pipeline backed by WebDataset tar shards.
+"""Load WebDataset shards into RF-DETR streaming datasets and loaders.
 
-A loose-file COCO split makes one ``open()`` per image per epoch. Raising ``num_workers`` to keep a GPU fed turns that
-into thousands of concurrent random opens, and the cost lands in the kernel rather than in decode work. Packing the
-same split into a few hundred ``.tar`` shards converts those random opens into sequential reads of large files: each
-worker walks its own shards front to back and pays one ``open()`` per shard instead of one per image.
-
-The module is deliberately split in two halves with different dependencies:
-
-- **Packing** (:func:`pack_coco_to_shards`) uses only the standard library's :mod:`tarfile`. A WebDataset shard is an
-  ordinary POSIX tar whose members are grouped by basename, so nothing about writing one needs the ``webdataset``
-  package. Keeping the packer dependency-free means it is exercised by the default CI job rather than only by an
-  extra-gated one.
-- **Reading** (:class:`WebDatasetDetection`, :func:`build_webdataset_loader`) imports ``webdataset`` lazily, so the
-  optional ``rfdetr[data]`` extra is only required by a run that actually streams shards.
-
-:func:`build_webdataset_loader` returns a stock :class:`~torch.utils.data.DataLoader` rather than
-``webdataset.WebLoader``; see that function for why the wrapper is dropped while the streaming pipeline is kept.
-
-Image bytes are copied into the shard verbatim — no re-encode — so a packed split decodes to the same pixels as the
-loose files it came from. Decoded samples are handed to :class:`~rfdetr.datasets.coco.ConvertCoco` and then to the
-transform pipeline built by :func:`~rfdetr.datasets.coco.make_coco_transforms`, which is the same CPU-side
-Albumentations/torchvision stack the loose-file path uses; batches leave the workers through the DataModule's own
-collate function, so the ``pin_memory`` hand-off and the Kornia GPU stage in
-:meth:`~rfdetr.training.module_data.RFDETRDataModule.on_after_batch_transfer` are untouched.
-
-Scope: detection and segmentation splits. Keypoint training is rejected explicitly — its label space is derived from a
-whole parsed COCO file (:func:`~rfdetr.datasets._keypoint_schema.infer_coco_keypoint_schema`), which a shard index
-does not carry.
-
-See https://github.com/roboflow/rf-detr/issues/1392.
+Purpose: Read indexed tar shards with RF-DETR transforms and deterministic
+distributed worker planning. Scope: the optional WebDataset dependency,
+iterable dataset, epoch sizing and DataLoader construction. Usage: build a
+dataset with build_webdataset and a loader with build_webdataset_loader.
+Outputs: RF-DETR-compatible dataset samples and DataLoaders. Failure: rejects
+missing indexes, incompatible distribution plans and unsupported keypoints.
+Used by: datasets, detr and training module data.
 """
 
 from __future__ import annotations
 
-import hashlib
 import io
 import json
 import random
-import shutil
-import tarfile
-import tempfile
-from collections import defaultdict
-from collections.abc import Callable, Iterator, Sequence
-from dataclasses import dataclass
+from collections.abc import Callable, Iterator
 from functools import partial
 from math import gcd
 from pathlib import Path, PurePath
-from typing import Any, Literal
+from typing import Any
 
 import numpy as np
 import torch
@@ -62,682 +34,35 @@ from torch.utils.data import DataLoader
 from rfdetr.datasets.coco import (
     ConvertCoco,
     draft_size_for_transforms,
-    filter_parent_categories,
     make_coco_transforms,
     make_coco_transforms_square_div_64,
     scale_coco_annotation,
 )
 from rfdetr.datasets.kornia_transforms import is_gpu_postprocess, resolve_backend_for_build
+from rfdetr.datasets.webdataset.index import (
+    WebDatasetSplitUnavailableError,
+    read_shard_index,
+)
+from rfdetr.datasets.webdataset.io import IMAGE_EXTENSIONS, resolve_within
 from rfdetr.utilities.logger import get_logger
 
 logger = get_logger()
-
-#: Schema version stamped into every shard index, so a future format change can be rejected with a clear message.
-INDEX_VERSION = 1
-
-#: Default shard size target in bytes (~100 MB), the size range WebDataset is tuned for.
-DEFAULT_MAX_SHARD_BYTES = 100 * 1024 * 1024
 
 #: Samples held in the training reservoir shuffle buffer, on top of shard-order shuffling.
 DEFAULT_SHUFFLE_BUFFER = 1000
 
 #: Warn once per loader when the smallest worker's shard share falls this far below the average.
-#:
-#: Shards split across workers by count, so an uneven split leaves the smallest worker short of the samples
-#: a fixed epoch asks of it: it wraps and repeats its own while better-supplied workers leave some unseen.
-#: Measured on a 2,000-image 3-epoch fine-tune, 39 shards over 8 workers (18% short) cost about 0.011 mAP@50:95
-#: against the map-style loader, while re-packing the same split into 290 shards (0.6% short) closed it.
 SHARD_SKEW_WARN_FRACTION = 0.05
 
 #: Raise instead of warn when the smallest worker's shard share falls this far below the average.
-#:
-#: Every degradation mode this module can measure is otherwise silent past a warning nobody watches live in a
-#: training log: a skew this severe is no longer a tuning nuisance worth a log line — it is close enough to a
-#: worker seeing a small fraction of its assigned share that failing the run before it starts is cheaper than
-#: discovering it from a degraded metric hours later. Must stay above :data:`SHARD_SKEW_WARN_FRACTION`.
 SHARD_SKEW_RAISE_FRACTION = 0.30
 
 #: Shards held in the training shard-order shuffle buffer.
-#:
-#: WebDataset rejects ``True`` here with a warning and silently substitutes this same value,
-#: so it is passed as an integer.
 DEFAULT_SHARD_SHUFFLE = 100
-
-CategoryIdPolicy = Literal["remap", "raw"]
-
-_IMAGE_EXTENSIONS: tuple[str, ...] = ("jpg", "jpeg", "png", "webp", "bmp")
-
-_TAR_BLOCK_BYTES = 512
-
-
-def _tar_member_bytes(payload_len: int) -> int:
-    """Bytes a POSIX ustar member of *payload_len* content bytes actually occupies in the archive.
-
-    Every member costs one 512-byte header block plus its content padded up to the next 512-byte boundary. Summing
-    raw payload lengths alone (as the packer's shard-size accounting used to) undercounts by this framing overhead
-    on every member — two per sample here (image, JSON sidecar) — which under-shoots ``max_shard_bytes`` and
-    produces shards measurably larger than requested, worse the smaller the average sample.
-
-    Args:
-        payload_len: Content length of the member, in bytes.
-
-    Returns:
-        Total bytes the member occupies in the tar archive, including its header and padding.
-
-    Examples:
-        >>> _tar_member_bytes(0)
-        512
-        >>> _tar_member_bytes(1)
-        1024
-        >>> _tar_member_bytes(512)
-        1024
-        >>> _tar_member_bytes(513)
-        1536
-    """
-    return _TAR_BLOCK_BYTES + -(-payload_len // _TAR_BLOCK_BYTES) * _TAR_BLOCK_BYTES
-
-
-_TAR_MEMBER_MODE = 0o644
-
-
-def _validate_split_name(split: str) -> str:
-    """Reject a split name that could escape the shard directory as a path component.
-
-    *split* becomes part of a shard or index file name with no other sanitisation, so a value containing a path
-    separator or a ``..`` segment would let :func:`_shard_name`/:func:`index_name` resolve outside the shard
-    directory — writing shards outside ``output_dir`` when packing, or reading an index from outside ``shard_dir``
-    when loading.
-
-    Args:
-        split: Split name supplied by a caller (CLI argument or library call).
-
-    Returns:
-        *split*, unchanged, once validated.
-
-    Raises:
-        ValueError: If *split* is empty, contains a path separator or a ``.``/``..`` segment, or carries a
-            glob metacharacter. The last one matters because the name is also matched against existing shard
-            files: a split called ``*`` would match — and then delete — the shards of every other split.
-
-    Examples:
-        >>> _validate_split_name("train")
-        'train'
-        >>> _validate_split_name("../escape")  # doctest: +ELLIPSIS
-        Traceback (most recent call last):
-            ...
-        ValueError: split '../escape' must not contain a path separator, drive prefix, or '..'.
-    """
-    if not split or "/" in split or "\\" in split or ":" in split or split in (".", ".."):
-        raise ValueError(f"split {split!r} must not contain a path separator, drive prefix, or '..'.")
-    forbidden = sorted(set("*?[]") & set(split))
-    if forbidden:
-        raise ValueError(
-            f"split {split!r} must not contain {''.join(forbidden)}: the name is matched against existing shard "
-            "files, so a glob metacharacter would reach other splits' shards."
-        )
-    return split
-
-
-def _shard_name(split: str, index: int, generation: str | None = None) -> str:
-    """Return the file name of shard *index* of *split*.
-
-    Args:
-        split: Split name the shard belongs to.
-        index: Zero-based shard number.
-        generation: Token identifying one pack run. Included in the name so that re-packing a split writes new
-            files instead of overwriting the ones the published index still points at.
-
-    Returns:
-        Shard file name.
-
-    Raises:
-        ValueError: If *split* would escape the shard directory (see :func:`_validate_split_name`).
-
-    Examples:
-        >>> _shard_name("train", 7)
-        'train-000007.tar'
-        >>> _shard_name("train", 7, "a1b2c3d4")
-        'train-a1b2c3d4-000007.tar'
-    """
-    _validate_split_name(split)
-    if generation is None:
-        return f"{split}-{index:06d}.tar"
-    return f"{split}-{generation}-{index:06d}.tar"
-
-
-def _resolve_within(base: Path, name: str) -> Path:
-    """Resolve *name* as a path component under *base*, rejecting any attempt to escape it.
-
-    Shard file names come from a shard index, which is ordinary on-disk JSON: a tampered or corrupted index can
-    carry an absolute path or a ``../``-laden entry. Two callers trust index entries without this check today —
-    opening a shard for reading and unlinking a stale shard during republish — and both would otherwise follow
-    the entry wherever it points.
-
-    Args:
-        base: Directory *name* must resolve inside.
-        name: Untrusted path-like string taken from a shard index.
-
-    Returns:
-        The resolved, absolute path to *name* under *base*.
-
-    Raises:
-        ValueError: If *name* is absolute or resolves outside *base*.
-
-    Examples:
-        >>> _resolve_within(Path("/data/shards"), "train-000000.tar").name
-        'train-000000.tar'
-        >>> _resolve_within(Path("/data/shards"), "../../etc/passwd")  # doctest: +ELLIPSIS
-        Traceback (most recent call last):
-            ...
-        ValueError: shard entry '../../etc/passwd' resolves outside ...shards.
-    """
-    base_resolved = base.resolve()
-    candidate = (base / name).resolve()
-    if not candidate.is_relative_to(base_resolved):
-        raise ValueError(f"shard entry {name!r} resolves outside {base}.")
-    return candidate
-
-
-def index_name(split: str) -> str:
-    """Return the file name of the shard index of *split*.
-
-    Args:
-        split: Split name the index describes.
-
-    Returns:
-        Index file name, relative to the shard directory.
-
-    Raises:
-        ValueError: If *split* would escape the shard directory (see :func:`_validate_split_name`).
-
-    Examples:
-        >>> index_name("val")
-        'val-index.json'
-    """
-    _validate_split_name(split)
-    return f"{split}-index.json"
-
-
-@dataclass(frozen=True)
-class ShardIndex:
-    """Everything a reader needs about one packed split without opening a shard.
-
-    Args:
-        split: Split name, matching the shard and index file names.
-        shards: Shard file names in packing order, relative to the index's directory.
-        num_samples: Total samples across all shards.
-        categories: COCO ``categories`` entries copied verbatim from the source annotation file.
-        annotated_category_ids: Category ids carrying at least one annotation in this split.
-        category_ids: ``"remap"`` when labels are contiguous 0-based indices, ``"raw"`` when they are the source
-            ``category_id`` values.
-        samples_per_shard: Sample count of each shard, aligned by position with *shards*. Shards are cut by byte
-            size, not sample count, so this can vary widely between shards of the same split; an empty tuple means
-            the index predates this field (or was built without it), and readers fall back to assuming a uniform
-            split.
-    """
-
-    split: str
-    shards: tuple[str, ...]
-    num_samples: int
-    categories: tuple[dict[str, Any], ...]
-    annotated_category_ids: tuple[int, ...]
-    category_ids: CategoryIdPolicy
-    samples_per_shard: tuple[int, ...] = ()
-
-    def to_json(self) -> dict[str, Any]:
-        """Return the JSON-serialisable form written to disk.
-
-        Returns:
-            Mapping stamped with :data:`INDEX_VERSION`.
-
-        Examples:
-            >>> index = ShardIndex("val", ("val-000000.tar",), 1, ({"id": 1, "name": "a"},), (1,), "remap", (1,))
-            >>> index.to_json()["num_samples"]
-            1
-        """
-        return {
-            "version": INDEX_VERSION,
-            "split": self.split,
-            "shards": list(self.shards),
-            "num_samples": self.num_samples,
-            "categories": list(self.categories),
-            "annotated_category_ids": list(self.annotated_category_ids),
-            "category_ids": self.category_ids,
-            "samples_per_shard": list(self.samples_per_shard),
-        }
-
-    @classmethod
-    def from_json(cls, payload: dict[str, Any]) -> ShardIndex:
-        """Rebuild an index from its on-disk mapping.
-
-        Args:
-            payload: Parsed index JSON.
-
-        Returns:
-            The reconstructed index.
-
-        Raises:
-            ValueError: If the index was written by an incompatible schema version.
-
-        Examples:
-            >>> payload = {
-            ...     "version": 1, "split": "val", "shards": ["val-000000.tar"], "num_samples": 1,
-            ...     "categories": [{"id": 1, "name": "a"}], "annotated_category_ids": [1], "category_ids": "remap",
-            ... }
-            >>> ShardIndex.from_json(payload).split
-            'val'
-        """
-        version = int(payload.get("version", 0))
-        if version != INDEX_VERSION:
-            raise ValueError(
-                f"Shard index schema version {version} is not supported by this RF-DETR release "
-                f"(expected {INDEX_VERSION}); re-pack the dataset with the current packer."
-            )
-        policy = payload.get("category_ids", "remap")
-        if policy not in ("remap", "raw"):
-            raise ValueError(f"Shard index declares unknown category_ids policy {policy!r}; expected 'remap' or 'raw'.")
-        return cls(
-            split=str(payload["split"]),
-            shards=tuple(str(shard) for shard in payload["shards"]),
-            num_samples=int(payload["num_samples"]),
-            categories=tuple(payload.get("categories", ())),
-            annotated_category_ids=tuple(int(cid) for cid in payload.get("annotated_category_ids", ())),
-            category_ids=policy,
-            samples_per_shard=tuple(int(count) for count in payload.get("samples_per_shard", ())),
-        )
-
-    def cat2label(self) -> dict[int, int] | None:
-        """Return the ``category_id`` to label-index mapping implied by this index.
-
-        Mirrors :class:`~rfdetr.datasets.coco.CocoDetection`'s detection branch: unannotated grouping categories are
-        dropped by :func:`~rfdetr.datasets.coco.filter_parent_categories` before indices are assigned, so a synthetic
-        Roboflow root category consumes no output slot.
-
-        Returns:
-            The mapping, or ``None`` when the index declares the ``"raw"`` policy and source ids are used as labels.
-
-        Examples:
-            >>> categories = ({"id": 3, "name": "a"}, {"id": 9, "name": "b"})
-            >>> ShardIndex("t", (), 0, categories, (3, 9), "remap").cat2label()
-            {3: 0, 9: 1}
-            >>> ShardIndex("t", (), 0, categories, (3, 9), "raw").cat2label() is None
-            True
-        """
-        if self.category_ids == "raw":
-            return None
-        kept = filter_parent_categories(list(self.categories), set(self.annotated_category_ids))
-        return {int(category["id"]): label for label, category in enumerate(kept)}
-
-
-class WebDatasetSplitUnavailableError(FileNotFoundError):
-    """Raised when a shard directory carries no index for the requested split.
-
-    A subclass of :class:`FileNotFoundError`, so callers that only care that the split is missing keep working, while
-    :meth:`~rfdetr.training.module_data.RFDETRDataModule._build_test_dataset` can tell "this split was never packed"
-    apart from any other missing file and fall back to ``val`` rather than aborting the run.
-    """
-
-
-def read_shard_index(shard_dir: str | Path, split: str) -> ShardIndex:
-    """Load the shard index of *split* from *shard_dir*.
-
-    Args:
-        shard_dir: Directory holding the shards and their index files.
-        split: Split name to read.
-
-    Returns:
-        The parsed index.
-
-    Raises:
-        WebDatasetSplitUnavailableError: If the split has no index file in *shard_dir*.
-    """
-    path = Path(shard_dir) / index_name(split)
-    if not path.exists():
-        raise WebDatasetSplitUnavailableError(
-            f"No WebDataset index for split {split!r} at {path}. "
-            f"Pack the split first: python -m rfdetr.cli.webdataset --split {split} ..."
-        )
-    with path.open(encoding="utf-8") as handle:
-        payload = json.load(handle)
-    index = ShardIndex.from_json(payload)
-    if index.split != split:
-        raise ValueError(
-            f"{path} is named for split {split!r} but its own recorded split is {index.split!r} — "
-            "the index file was likely copied or renamed rather than produced by packing this split; "
-            "repack it instead of renaming an existing index."
-        )
-    return index
-
-
-# ----------------------------------------------------------------------------------------------------------------
-# Packing
-# ----------------------------------------------------------------------------------------------------------------
-
-
-def _add_bytes(tar: tarfile.TarFile, name: str, payload: bytes) -> None:
-    """Append *payload* to *tar* as a regular member named *name*.
-
-    Ownership and timestamps are zeroed so that packing the same split twice produces byte-identical shards.
-
-    Args:
-        tar: Open archive to append to.
-        name: Member name, whose extension becomes the sample's WebDataset field.
-        payload: Member contents.
-    """
-    info = tarfile.TarInfo(name)
-    info.size = len(payload)
-    info.mode = _TAR_MEMBER_MODE
-    info.mtime = 0
-    info.uid = 0
-    info.gid = 0
-    info.uname = ""
-    info.gname = ""
-    tar.addfile(info, io.BytesIO(payload))
-
-
-def _annotations_by_image(coco_data: dict[str, Any]) -> dict[Any, list[dict[str, Any]]]:
-    """Group a parsed COCO file's annotations by ``image_id``.
-
-    Args:
-        coco_data: Parsed COCO JSON.
-
-    Returns:
-        Mapping from image id to its annotation list; images without annotations are absent.
-
-    Examples:
-        >>> _annotations_by_image({"annotations": [{"image_id": 5, "category_id": 1}]})[5][0]["category_id"]
-        1
-    """
-    grouped: dict[Any, list[dict[str, Any]]] = defaultdict(list)
-    for annotation in coco_data.get("annotations", []):
-        grouped[annotation["image_id"]].append(annotation)
-    return grouped
-
-
-def _shard_content_digest(path: Path) -> str:
-    """Return a streamed SHA-256 hex digest of *path*'s bytes.
-
-    Args:
-        path: Shard file to digest.
-
-    Returns:
-        Full 64-character hex digest.
-    """
-    hasher = hashlib.sha256()
-    with path.open("rb") as handle:
-        for chunk in iter(lambda: handle.read(1024 * 1024), b""):
-            hasher.update(chunk)
-    return hasher.hexdigest()
-
-
-def _pack_generation(payload: dict[str, Any], shard_digests: list[str]) -> str:
-    """Derive a short generation token from what was packed.
-
-    Deterministic on purpose: the same split packed twice produces the same token, so re-packing unchanged data
-    reproduces the same file names rather than churning the directory. Any change to the sample count, the
-    category set, the per-shard sample counts or a shard's actual bytes changes the token, which is what keeps
-    a re-pack from writing over the shards a currently published index still references — including the case a
-    byte size alone would miss: a changed image or annotation whose re-encoded shard happens to land on the same
-    padded tar size as the one it replaces.
-
-    Args:
-        payload: The index mapping for this pack, excluding the shard names it is about to name.
-        shard_digests: SHA-256 hex digest of each staged shard's content, in order (see
-            :func:`_shard_content_digest`).
-
-    Returns:
-        Eight hex characters.
-
-    Examples:
-        >>> a = _pack_generation({"split": "train", "num_samples": 2}, ["aa", "bb"])
-        >>> a == _pack_generation({"split": "train", "num_samples": 2}, ["aa", "bb"])
-        True
-        >>> a == _pack_generation({"split": "train", "num_samples": 2}, ["aa", "cc"])
-        False
-    """
-    material = json.dumps({"index": payload, "digests": shard_digests}, sort_keys=True).encode("utf-8")
-    return hashlib.sha256(material).hexdigest()[:8]
-
-
-def _published_shard_names(destination: Path, split: str) -> set[str]:
-    """Return the shard names the currently published index of *split* references.
-
-    Reading the index rather than globbing keeps cleanup scoped to what this split actually published: the
-    directory legitimately holds other splits, and a half-written generation from a crashed run must not be
-    mistaken for a live one.
-
-    Args:
-        destination: Directory holding the pack.
-        split: Split whose published index to read.
-
-    Returns:
-        Shard file names from the published index, or an empty set when the split has no readable index yet.
-
-    Examples:
-        >>> _published_shard_names(Path("/nonexistent"), "train")
-        set()
-    """
-    try:
-        return set(read_shard_index(destination, split).shards)
-    except (WebDatasetSplitUnavailableError, ValueError, KeyError, json.JSONDecodeError):
-        return set()
-
-
-def pack_coco_to_shards(
-    image_dir: str | Path,
-    annotations_file: str | Path,
-    output_dir: str | Path,
-    *,
-    split: str = "train",
-    max_shard_bytes: int = DEFAULT_MAX_SHARD_BYTES,
-    category_ids: CategoryIdPolicy = "remap",
-) -> ShardIndex:
-    """Pack a COCO-format split into WebDataset tar shards plus a JSON index.
-
-    Every image becomes two adjacent tar members sharing one basename: the original image file copied byte for byte,
-    and a ``.json`` sidecar holding that image's ``image_id``, ``file_name`` and annotation list. A new shard is
-    started once the current one reaches *max_shard_bytes*, so shards land near that size rather than exactly on it —
-    a sample is never split across two shards. Size accounting counts each member's actual tar footprint (its
-    512-byte header plus content padded to the next 512-byte boundary, see :func:`_tar_member_bytes`), not just raw
-    payload length, so shards land close to *max_shard_bytes* even with many small samples.
-
-    Args:
-        image_dir: Directory holding the split's image files, named as the annotation file's ``file_name`` entries.
-        annotations_file: COCO-format JSON annotation file for the split.
-        output_dir: Directory to write shards and the index into; created if absent.
-        split: Split name used for the shard and index file names.
-        max_shard_bytes: Size at which the current shard is closed and the next one opened.
-        category_ids: ``"remap"`` assigns contiguous 0-based labels, matching
-            :func:`~rfdetr.datasets.coco.build_roboflow_from_coco`. ``"raw"`` keeps the source ``category_id`` values
-            as labels, matching :func:`~rfdetr.datasets.coco.build_coco`'s convention for evaluating a model trained
-            on the COCO-2017 label space.
-
-    Returns:
-        The index describing what was written.
-
-    Raises:
-        FileNotFoundError: If *annotations_file* or a listed image file does not exist. A missing image is fatal
-            rather than skipped: dropping training images silently is worse than failing on the file that is absent.
-        ValueError: If *max_shard_bytes* is not positive, if the annotation file lists no images, if an image's
-            ``file_name`` resolves outside *image_dir*, if an image has an extension the reader cannot decode, or
-            if any annotation's ``image_id`` matches no image — the shape an ``id``/``image_id`` type mismatch
-            takes, which would otherwise pack a split that silently drops or mislabels samples. Packing stops
-            before writing anything to *output_dir* on any of these failures, so a re-pack that fails partway
-            through never leaves a previously-packed, valid split there in a corrupted, half-overwritten state.
-    """
-    if max_shard_bytes <= 0:
-        raise ValueError(f"max_shard_bytes must be > 0, got {max_shard_bytes}.")
-    if category_ids not in ("remap", "raw"):
-        # The CLI constrains this with `choices`, but a library caller can pass anything. Without this the pack
-        # succeeds and only fails later, at read time, when ShardIndex.from_json rejects the policy it wrote.
-        raise ValueError(f"category_ids must be 'remap' or 'raw', got {category_ids!r}.")
-    _validate_split_name(split)
-
-    image_root = Path(image_dir)
-    annotations_path = Path(annotations_file)
-    if not annotations_path.exists():
-        raise FileNotFoundError(f"COCO annotation file {annotations_path} does not exist.")
-
-    with annotations_path.open(encoding="utf-8") as handle:
-        coco_data: dict[str, Any] = json.load(handle)
-
-    images: list[dict[str, Any]] = list(coco_data.get("images", []))
-    if not images:
-        raise ValueError(f"COCO annotation file {annotations_path} lists no images; nothing to pack.")
-
-    grouped = _annotations_by_image(coco_data)
-    image_ids = {image_entry["id"] for image_entry in images}
-    orphaned_ids = sorted((set(grouped) - image_ids), key=str)
-    if orphaned_ids:
-        raise ValueError(
-            f"{annotations_path} has {len(orphaned_ids)} annotation image_id value(s) matching no image "
-            f"(e.g. {orphaned_ids[0]!r}): those annotations would be dropped silently. The usual cause is a type "
-            "mismatch between images[].id and annotations[].image_id (one string, one int)."
-        )
-
-    destination = Path(output_dir)
-    destination.mkdir(parents=True, exist_ok=True)
-    work_dir = Path(tempfile.mkdtemp(prefix=f".{split}-pack-", dir=destination))
-
-    shard_names: list[str] = []
-    shard_sample_counts: list[int] = []
-    annotated_ids: set[int] = set()
-    shard_index = 0
-    written = 0
-    shard_samples = 0
-    tar: tarfile.TarFile | None = None
-    shard_bytes = 0
-
-    try:
-        for position, image_entry in enumerate(images):
-            file_name = str(image_entry["file_name"])
-            candidate = PurePath(file_name)
-            if candidate.anchor or ".." in candidate.parts:
-                # Lexical check only, deliberately not `.resolve()` + `is_relative_to`: `.resolve()` follows
-                # symlinks, and a legitimately symlinked image tree (the standard way a large COCO-scale split is
-                # shared on a cluster without duplicating storage) would then resolve outside `image_root` and be
-                # rejected even though `file_name` itself never leaves it. The actual threat is a crafted
-                # `file_name` string, which is lexical by construction.
-                raise ValueError(
-                    f"Image file_name {file_name!r} in {annotations_path} is absolute or escapes {image_root} "
-                    "via '..'; refusing to read a file the split's image directory does not own."
-                )
-            source = image_root / file_name
-            if not source.exists():
-                raise FileNotFoundError(
-                    f"Image {source} listed in {annotations_path} does not exist; "
-                    "packing stops rather than silently dropping it from the training set."
-                )
-            extension = source.suffix.lower().lstrip(".")
-            if extension not in _IMAGE_EXTENSIONS:
-                raise ValueError(
-                    f"Image {source} has extension {extension or '(none)'}, which the shard reader cannot decode "
-                    f"(supported: {', '.join(_IMAGE_EXTENSIONS)}). Convert the split before packing, so this fails "
-                    "here rather than on the first training batch."
-                )
-            payload = source.read_bytes()
-            annotations = grouped.get(image_entry["id"], [])
-            annotated_ids.update(int(annotation["category_id"]) for annotation in annotations)
-            sidecar = json.dumps(
-                {
-                    "image_id": image_entry["id"],
-                    "file_name": image_entry["file_name"],
-                    "annotations": annotations,
-                }
-            ).encode("utf-8")
-
-            if tar is None:
-                name = _shard_name(split, shard_index)
-                tar = tarfile.open(work_dir / name, "w")
-                shard_names.append(name)
-                shard_bytes = 0
-                shard_samples = 0
-
-            key = f"{position:08d}"
-            _add_bytes(tar, f"{key}.{extension}", payload)
-            _add_bytes(tar, f"{key}.json", sidecar)
-            shard_bytes += _tar_member_bytes(len(payload)) + _tar_member_bytes(len(sidecar))
-            written += 1
-            shard_samples += 1
-
-            if shard_bytes >= max_shard_bytes:
-                tar.close()
-                tar = None
-                shard_sample_counts.append(shard_samples)
-                shard_index += 1
-
-        if tar is not None:
-            tar.close()
-            tar = None
-            shard_sample_counts.append(shard_samples)
-
-        provisional = list(shard_names)
-        shard_digests = [_shard_content_digest(work_dir / name) for name in provisional]
-        generation = _pack_generation(
-            {
-                "split": split,
-                "num_samples": written,
-                "categories": list(coco_data.get("categories", ())),
-                "annotated_category_ids": sorted(annotated_ids),
-                "category_ids": category_ids,
-                "samples_per_shard": shard_sample_counts,
-            },
-            shard_digests,
-        )
-        shard_names = [_shard_name(split, position, generation) for position in range(len(provisional))]
-        index = ShardIndex(
-            split=split,
-            shards=tuple(shard_names),
-            num_samples=written,
-            categories=tuple(coco_data.get("categories", ())),
-            annotated_category_ids=tuple(sorted(annotated_ids)),
-            category_ids=category_ids,
-            samples_per_shard=tuple(shard_sample_counts),
-        )
-        index_bytes = json.dumps(index.to_json()).encode("utf-8")
-
-        # Every shard packed successfully: publish shards and the index together. A previous pack's shards that
-        # this run does not reproduce (e.g. it produced fewer of them) are removed too, so no stale, unindexed
-        # shard is left behind. The index is written last, so a reader can never observe an index whose shard
-        # list is only partially on disk.
-        # Shard names carry this run's generation token (content-derived, see _pack_generation), so moving them
-        # in cannot overwrite a shard the published index still points at. Publication is therefore: move this
-        # generation's shards in, swap the index with os.replace (atomic on POSIX), then drop the shards only
-        # the previous index referenced. A reader that opened the old index keeps a complete pack until that
-        # last step; one that opens the new index sees this generation in full.
-        # This narrows, but does not close, the window where a concurrent reader loses files mid-epoch: a
-        # reader that opened the old index and has not yet finished reading every shard it names can still hit
-        # a missing file once the cleanup step below runs. Closing that fully needs reference-counted or
-        # TTL-based garbage collection of old generations across active readers, which is a design decision
-        # beyond what this module tracks today (no reader registry exists to know when it is safe to delete).
-        previous = _published_shard_names(destination, split)
-        for staged, published in zip(provisional, shard_names):
-            (work_dir / staged).replace(destination / published)
-        # Staged inside work_dir, not destination: work_dir was created with dir=destination (same filesystem,
-        # so the replace() below stays an atomic rename), and this way a failure between write_bytes and
-        # replace() leaves the partial file inside work_dir, where the finally block's rmtree cleans it up
-        # instead of leaving a stray dotfile behind in destination.
-        staged_index = work_dir / f".{index_name(split)}.{generation}"
-        staged_index.write_bytes(index_bytes)
-        staged_index.replace(destination / index_name(split))
-        for name in previous - set(shard_names):
-            _resolve_within(destination, name).unlink(missing_ok=True)
-    finally:
-        if tar is not None:
-            tar.close()
-        shutil.rmtree(work_dir, ignore_errors=True)
-
-    logger.info("Packed %d %s samples into %d shard(s) at %s", written, split, len(shard_names), destination)
-    return index
-
-
-# ----------------------------------------------------------------------------------------------------------------
-# Reading
-# ----------------------------------------------------------------------------------------------------------------
 
 
 def _require_webdataset() -> Any:
-    """Import and return the ``webdataset`` module, or raise an actionable ``ImportError``."""
+    """Import and return the WebDataset module, or raise an actionable ImportError."""
     try:
         import webdataset
     except ImportError as exc:  # pragma: no cover - exercised only without the optional extra
@@ -1008,12 +333,12 @@ class WebDatasetDetection(torch.utils.data.IterableDataset[tuple[Any, Any]]):
         Raises:
             KeyError: If the sample carries no recognised image member or no ``json`` sidecar.
         """
-        extension = next((candidate for candidate in _IMAGE_EXTENSIONS if candidate in sample), None)
+        extension = next((candidate for candidate in IMAGE_EXTENSIONS if candidate in sample), None)
         if extension is None:
             present = sorted(key for key in sample if not key.startswith("__"))
             raise KeyError(
                 f"WebDataset sample {sample.get('__key__', '?')} has no image member "
-                f"(looked for {', '.join(_IMAGE_EXTENSIONS)}); found {present}."
+                f"(looked for {', '.join(IMAGE_EXTENSIONS)}); found {present}."
             )
         if "json" not in sample:
             raise KeyError(f"WebDataset sample {sample.get('__key__', '?')} has no 'json' annotation sidecar.")
@@ -1038,7 +363,7 @@ class WebDatasetDetection(torch.utils.data.IterableDataset[tuple[Any, Any]]):
         """Return this split's shards as ``file:`` URLs.
 
         See :func:`_shard_url` for why a shard is named as a ``file:`` URL rather than by its plain path, and
-        :func:`_resolve_within` for why each entry is validated against ``self._shard_dir`` before being opened —
+        :func:`resolve_within` for why each entry is validated against ``self._shard_dir`` before being opened —
         the index is on-disk JSON, not a trusted value.
 
         Returns:
@@ -1047,7 +372,7 @@ class WebDatasetDetection(torch.utils.data.IterableDataset[tuple[Any, Any]]):
         Raises:
             ValueError: If a shard entry is absolute or resolves outside ``self._shard_dir``.
         """
-        return [_shard_url(_resolve_within(self._shard_dir, shard)) for shard in self.index.shards]
+        return [_shard_url(resolve_within(self._shard_dir, shard)) for shard in self.index.shards]
 
     def __iter__(self) -> Iterator[tuple[Any, Any]]:
         """Iterate this worker's share of the split.
@@ -1483,27 +808,3 @@ def build_webdataset(image_set: str, args: Any, resolution: int) -> WebDatasetDe
         # Evaluation indexes may omit categories; names must follow the same train-owned label space as labels.
         dataset._label_categories = train_index.categories
     return dataset
-
-
-# ----------------------------------------------------------------------------------------------------------------
-# Compatibility CLI entry point
-# ----------------------------------------------------------------------------------------------------------------
-
-
-def main(argv: Sequence[str] | None = None) -> int:
-    """Preserve the original packing entry point; see :mod:`rfdetr.cli.webdataset`.
-
-    Args:
-        argv: Argument list, or ``None`` to read ``sys.argv``.
-
-    Returns:
-        Process exit status.
-    """
-    # Import on invocation to avoid a cycle: the CLI imports the reusable IO functions.
-    from rfdetr.cli.webdataset import main as cli_main
-
-    return cli_main(argv)
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
